Polish sanitizer: dedicated test file, canonical docstring enumeration
